+++ ui/src/flux/components/TimeMachineTables.tsx
@@ -1,17 +1,57 @@
+// Libraries
 import React, {PureComponent} from 'react'
+import memoizeOne from 'memoize-one'
 
+// Components
 import TableSidebar from 'src/flux/components/TableSidebar'
 import {FluxTable} from 'src/types'
 import NoResults from 'src/flux/components/NoResults'
-import TimeMachineTable from 'src/flux/components/TimeMachineTable'
+import TableGraph from 'src/shared/components/TableGraph'
+
+// Utils
+import {getDeep} from 'src/utils/wrappers'
+
+// Types
+import {QueryUpdateState} from 'src/shared/actions/queries'
+import {ColorString} from 'src/types/colors'
+import {TableOptions, FieldOption, DecimalPlaces} from 'src/types/dashboards'
+import {DataType} from 'src/shared/constants'
 
 interface Props {
   data: FluxTable[]
+  dataType: DataType
+  tableOptions: TableOptions
+  timeFormat: string
+  decimalPlaces: DecimalPlaces
+  fieldOptions: FieldOption[]
+  handleSetHoverTime?: (hovertime: string) => void
+  colors: ColorString[]
+  editorLocation?: QueryUpdateState
 }
 
 interface State {
   selectedResultID: string | null
 }
+
+const filterTables = (tables: FluxTable[]): FluxTable[] => {
+  const IGNORED_COLUMNS = ['', 'result', 'table', '_start', '_stop']
+
+  return tables.map(table => {
+    const header = table.data[0]
+    const indices = IGNORED_COLUMNS.map(name => header.indexOf(name))
+    const tableData = table.data
+    const data = tableData.map(row => {
+      return row.filter((__, i) => !indices.includes(i))
+    })
+
+    return {
+      ...table,
+      data,
+    }
+  })
+}
+
+const filteredTablesMemoized = memoizeOne(filterTables)
 
 class TimeMachineTables extends PureComponent<Props, State> {
   constructor(props) {
@@ -29,6 +69,15 @@ class TimeMachineTables extends PureComponent<Props, State> {
   }
 
   public render() {
+    const {
+      colors,
+      dataType,
+      timeFormat,
+      tableOptions,
+      decimalPlaces,
+      editorLocation,
+      handleSetHoverTime,
+    } = this.props
     return (
       <div className="time-machine-tables">
         {this.showSidebar && (
@@ -39,11 +88,42 @@ class TimeMachineTables extends PureComponent<Props, State> {
           />
         )}
         {this.shouldShowTable && (
-          <TimeMachineTable table={this.selectedResult} />
+          <TableGraph
+            data={this.selectedResult}
+            dataType={dataType}
+            colors={colors}
+            tableOptions={tableOptions}
+            fieldOptions={this.fieldOptions}
+            timeFormat={timeFormat}
+            decimalPlaces={decimalPlaces}
+            editorLocation={editorLocation}
+            handleSetHoverTime={handleSetHoverTime}
+          />
         )}
         {!this.hasResults && <NoResults />}
       </div>
     )
+  }
+
+  private get fieldOptions(): FieldOption[] {
+    const {fieldOptions} = this.props
+
+    const internalName = getDeep<string>(fieldOptions, '0.internalName', '')
+    if (
+      fieldOptions.length === 0 ||
+      (fieldOptions.length === 1 &&
+        (internalName === 'time' || internalName === '_time'))
+    ) {
+      const headers = getDeep(this.selectedResult, 'data.0', [])
+
+      return headers.map(h => ({
+        internalName: h,
+        displayName: '',
+        visible: true,
+      }))
+    }
+
+    return fieldOptions
   }
 
   private handleSelectResult = (selectedResultID: string): void => {
@@ -73,7 +153,9 @@ class TimeMachineTables extends PureComponent<Props, State> {
   }
 
   private get selectedResult(): FluxTable {
-    return this.props.data.find(d => d.id === this.state.selectedResultID)
+    const filteredTables = filteredTablesMemoized(this.props.data)
+
+    return filteredTables.find(d => d.id === this.state.selectedResultID)
   }
 }