+++ ui/src/flux/components/YieldNodeVis.tsx
@@ -7,10 +7,15 @@ import {Radio} from 'src/reusable_ui'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
 import {FluxTable} from 'src/types'
+import {VisualizationOptions} from 'src/types/dataExplorer'
+import {DataType} from 'src/shared/constants'
+import {getCellTypeColors} from 'src/dashboards/constants/cellEditor'
+import {CellType} from 'src/types/dashboards'
 
 interface Props {
   data: FluxTable[]
   yieldName: string
+  visualizationOptions: VisualizationOptions
 }
 
 enum VisType {
@@ -66,7 +71,17 @@ class YieldNodeVis extends PureComponent<Props, State> {
 
   private get vis(): JSX.Element {
     const {visType} = this.state
-    const {data} = this.props
+    const {data, visualizationOptions} = this.props
+
+    const {
+      tableOptions,
+      timeFormat,
+      decimalPlaces,
+      fieldOptions,
+      thresholdsListColors,
+      gaugeColors,
+      lineColors,
+    } = visualizationOptions
 
     if (visType === VisType.Line) {
       return (
@@ -76,7 +91,24 @@ class YieldNodeVis extends PureComponent<Props, State> {
       )
     }
 
-    return <TimeMachineTables data={data} />
+    const colors = getCellTypeColors({
+      cellType: CellType.Table,
+      gaugeColors,
+      thresholdsListColors,
+      lineColors,
+    })
+
+    return (
+      <TimeMachineTables
+        data={data}
+        dataType={DataType.flux}
+        tableOptions={tableOptions}
+        timeFormat={timeFormat}
+        decimalPlaces={decimalPlaces}
+        fieldOptions={fieldOptions}
+        colors={colors}
+      />
+    )
   }
 
   private selectVisType = (visType: VisType): void => {